micro-service-gradle/micro-service-a/src/main/java/cn/plantlink/a/service/AService.java
package cn.plantlink.a.service;

import org.springframework.boot.context.properties.EnableConfigurationProperties;
import org.springframework.stereotype.Service;

@Service
@EnableConfigurationProperties(ServiceProperties.class)
public class AService {

    private final ServiceProperties serviceProperties;

    public AService(ServiceProperties serviceProperties) {
        this.serviceProperties = serviceProperties;
    }

    public String message() {
        return this.serviceProperties.getMessage();
    }
}